Implement user right refs #263
Merge of branch add_acl

Squashed commit of the following:

commit a3265a71014fc74bcc456aa9436494fc4ed434ea
Merge: 2522634 b231b14
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Tue Jun 30 09:58:23 2015 +0200

    Merge remote-tracking branch 'origin/master' into add_acl

    in order to future merging of add_acl into master

    Conflicts:
    	Tests/Controller/PersonControllerCreateTest.php

commit 252263400b1fb54091a0cd36d9ab84812c3a9942
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Mon Jun 29 23:00:25 2015 +0200

    check acl on timeline page

commit 5f08806585ce4ce1d58c6c507b191221b924d861
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 19 22:24:18 2015 +0200

    remove dead code

commit 860d62d9c0d9db2186eee725e7ef553ddf1f98f3
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 19 22:19:02 2015 +0200

    add access condition for link to person view in person view

commit e7b295000dd420309288e83920f6656679d36173
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 19 22:13:04 2015 +0200

    add ACL on person view

commit 7b1f90a2a28c1aec5ee0eb78206c75a9f1e09c60
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 19 21:47:49 2015 +0200

    add autorization for filtering person results in search

commit c6669164e2e36b555bed0745d5c7c12b53468b97
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 19 21:47:26 2015 +0200

    add fixtures for roles defined by person bundle

commit 8f30f5222efe531fb77a2e22b6af8069c5814d00
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 19 20:13:17 2015 +0200

    [transfer purpose][ci-skip]

commit e8ca4ce4842e619342a90d4343b725c2946f7d70
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 19 17:31:27 2015 +0200

    should fix reference problem

commit 8c32524a0fd61d504e9640705b97d10bd9b2b993
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 19 17:29:48 2015 +0200

    fix test about inheritance

commit d93cef2c596aba7fc977cc8db354241d6a77e6e2
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Thu Jun 18 21:38:59 2015 +0200

    add role hierarchy + test

commit 5066cc5318d67f20e3cff9f91f6d2b6a94d5f9c1
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Thu Jun 18 01:28:40 2015 +0200

    add tests for person voter

commit cf3359b11b16972db3c55102a3c8b2aae2c27789
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Thu Jun 18 01:07:30 2015 +0200

    fix "pass by reference" error

commit 94942eb34330813a10124ba58ec814ca685930c5
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Wed Jun 17 22:04:49 2015 +0200

    fix syntax error ($this->$is not correct)

commit 82d25df6b0e21b4032f4fc82dee203ccd1a067dd
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Wed Jun 17 21:51:41 2015 +0200

    add first test for person voter

commit 84e0d05be07477175db1ef95f3716a3c420291ce
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Wed Jun 17 21:50:06 2015 +0200

    define person voter as a service + CHILL_PERSON_SEE

commit bd55e505a179b0df4133903f766d9a8feb469900
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Tue Jun 16 20:57:25 2015 +0200

    add test for creating person with user multi center

commit d29147c8e70b2ae8313d973fe4b3d45d3f064d9b
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Mon Jun 15 09:11:56 2015 +0200

    fix tests - add center to persons

    since a center is associated with center, people used in tests must be associated with a center.
    => associate a center to people used in tests

commit cfac1ea68aaf615b5f8e9e67388f5407e5095d72
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Mon Jun 15 09:06:24 2015 +0200

    remove errors on deprecation

commit e1df87df87e831f3827721313603bcf41b3da891
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 12 22:31:03 2015 +0200

    fix migration error

commit 733fa96a2d874b84d2a0ece6664e4a64eed10b4d
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Wed Jun 10 21:17:15 2015 +0200

    add center type on creation forms

    tests are not adapted yet [ci skip]

commit c0d1d1b1c8fd58d9c15bf59d66de28a848df6d35
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Tue Jun 9 23:11:07 2015 +0200

    fix bug in migration

commit 693abfb7253dfe955beb63de3aebc05a15363cdb
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Tue Jun 9 22:41:21 2015 +0200

    add checks on migration

    check that there are data or not in database and, if there are data,
    create default center if required.

    If not required, the default center is the first in the list.

commit 8127ea8d49e094c81ce65e838416953bc8d47678
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Mon Jun 8 00:33:39 2015 +0200

    add first impl of PersonVoter (not tested) [ci skip]

commit 2ea674dbcf8848603f0fefa109abf2412722013f
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Mon Jun 8 00:32:46 2015 +0200

    add center to person

commit 0b1e3f137ace9eeb959b80cdbdfa3636c75af715
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Sat Jun 6 11:02:49 2015 +0200

    add phpunit bridge to handle deprecation msg

commit 70a4fd16181badad8ac9b87fd24e1000c7b1a734
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 5 12:42:01 2015 +0200

    remove deprecdation warning in phpunit

commit b3b8d1deefbdc72505606df7ca217771277e3555
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 5 11:41:36 2015 +0200

    fix deprecation of pattern in routing

    pattern => path in routing.yml

commit eea1252d8517ee582c958dc6045f05c1438dd9d1
Author: Julien Fastré <julien.fastre@champs-libres.coop>
Date:   Fri Jun 5 10:56:02 2015 +0200

    switch to symfony 2.7 [ci-skip]

+++ Tests/Controller/PersonControllerCreateTest.php
@@ -36,6 +36,8 @@ class PersonControllerCreateTest extends WebTestCase
     const GENRE_INPUT = "chill_personbundle_person_creation[genre]";
     const DATEOFBIRTH_INPUT = "chill_personbundle_person_creation[dateOfBirth]";
     const CREATEDATE_INPUT = "chill_personbundle_person_creation[creation_date]";
+    const CENTER_INPUT = "chill_personbundle_person_creation[center]";
+    
     const LONG_TEXT = "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Integer nec odio. Praesent libero. Sed cursus ante dapibus diam. Sed nisi. Nulla quis sem at nibh elementum imperdiet. Duis sagittis ipsum. Praesent mauris. Fusce nec tellus sed augue semper porta. Mauris massa. Vestibulum lacinia arcu eget nulla. Class aptent taciti sociosq.";
     
     /**
@@ -43,10 +45,10 @@ class PersonControllerCreateTest extends WebTestCase
      * 
      * @return \Symfony\Component\BrowserKit\Client
      */
-    private function getAuthenticatedClient()
+    private function getAuthenticatedClient($username = 'center a_social')
     {
         return static::createClient(array(), array(
-           'PHP_AUTH_USER' => 'center a_social',
+           'PHP_AUTH_USER' => $username,
            'PHP_AUTH_PW'   => 'password',
         ));
     }
@@ -55,10 +57,11 @@ class PersonControllerCreateTest extends WebTestCase
      * 
      * @param Form $creationForm
      */
-    private function fillAValidCreationForm(Form &$creationForm)
+    private function fillAValidCreationForm(Form &$creationForm, 
+          $firstname = 'God', $lastname = 'Jesus')
     {
-        $creationForm->get(self::FIRSTNAME_INPUT)->setValue("God");
-        $creationForm->get(self::LASTNAME_INPUT)->setValue("Jesus");
+        $creationForm->get(self::FIRSTNAME_INPUT)->setValue($firstname);
+        $creationForm->get(self::LASTNAME_INPUT)->setValue($lastname);
         $creationForm->get(self::GENRE_INPUT)->select("man");
         $date = new \DateTime('1947-02-01');
         $creationForm->get(self::DATEOFBIRTH_INPUT)->setValue($date->format('d-m-Y'));
@@ -203,13 +206,57 @@ class PersonControllerCreateTest extends WebTestCase
             . "/{_locale}/person/{personID}/general");
     }
     
+    /**
+     * test adding a person with a user with multi center
+     * is valid
+     */
+    public function testValidFormWithMultiCenterUser()
+    {
+        $client = $this->getAuthenticatedClient('multi_center');
+        
+        $crawler = $client->request('GET', '/fr/person/new');
+        
+        $this->assertTrue($client->getResponse()->isSuccessful(), 
+              "The page is accessible at the URL /{_locale}/person/new");
+        $form = $crawler->selectButton("Ajouter la personne")->form();
+        
+        $this->fillAValidCreationForm($form, 'roger', 'rabbit');
+        
+        $this->assertTrue($form->has(self::CENTER_INPUT),
+              'The page contains a "center" input');
+        $centerInput = $form->get(self::CENTER_INPUT);
+        $availableValues = $centerInput->availableOptionValues();
+        $lastCenterInputValue = end($availableValues);
+        $centerInput->setValue($lastCenterInputValue);
+        
+        $client->submit($form);
+        
+        $this->assertTrue($client->getResponse()->isRedirect(), 
+              "a valid form redirect to url /{_locale}/person/{personId}/general/edit");
+        $client->followRedirect();
+        $this->assertRegExp('|/fr/person/[1-9][0-9]*/general/edit$|', 
+              $client->getHistory()->current()->getUri(), 
+              "a valid form redirect to url /{_locale}/person/{personId}/general/edit");
+             
+    }
+    
     public static function tearDownAfterClass()
     {
         static::bootKernel();
         $em = static::$kernel->getContainer()->get('doctrine.orm.entity_manager');
+        
+        //remove two people created during test
         $jesus = $em->getRepository('ChillPersonBundle:Person')
               ->findOneBy(array('firstName' => 'God'));
-        $em->remove($jesus);
+        if ($jesus !== NULL) {
+            $em->remove($jesus);
+        }
+        
+        $jesus2 = $em->getRepository('ChillPersonBundle:Person')
+              ->findOneBy(array('firstName' => 'roger'));
+        if ($jesus2 !== NULL) {
+            $em->remove($jesus2);
+        }
         $em->flush();
     }
 }
